+++ src/vs/sessions/contrib/agentFeedback/browser/agentFeedbackItemsBackend.ts
@@ -17,7 +17,7 @@ import { ICodeReviewSuggestion } from '../../codeReview/browser/codeReviewServic
 import { IAgentHostSessionsProvider, isAgentHostProviderId } from '../../../common/agentHostSessionsProvider.js';
 import { ISessionsManagementService } from '../../../services/sessions/common/sessionsManagement.js';
 import { ISessionsProvidersService } from '../../../services/sessions/browser/sessionsProvidersService.js';
-import { AgentFeedbackKind, AgentFeedbackState, IAgentFeedback } from './agentFeedbackService.js';
+import { AgentFeedbackKind, AgentFeedbackState, IAgentFeedback } from './agentFeedbackModel.js';
 
 // --- Backend interface --------------------------------------------------------
